Move registry code for smaller diff
