+++ src/network/mod.rs
@@ -2,4 +2,3 @@ pub mod api;
 pub mod management;
 pub mod raft;
 pub mod raft_network_impl;
-pub mod registry;
