+++ src/registry/blobs/uploads/patch.rs
@@ -1,7 +1,7 @@
 use crate::app::RegistryApp;
 use crate::extractors::Token;
-use crate::network::registry::errors::RegistryError;
-use crate::network::registry::utils::upload_part;
+use crate::registry::errors::RegistryError;
+use crate::registry::utils::upload_part;
 use crate::types::RepositoryName;
 use actix_web::http::StatusCode;
 use actix_web::web::Payload;
