+++ src/registry/manifests/get.rs
@@ -1,6 +1,6 @@
 use crate::app::RegistryApp;
 use crate::extractors::Token;
-use crate::network::registry::errors::RegistryError;
+use crate::registry::errors::RegistryError;
 use crate::types::Digest;
 use crate::types::RepositoryName;
 use actix_files::NamedFile;
