+++ src/registry/mod.rs
