+++ src/registry/tags/get.rs
@@ -1,6 +1,6 @@
 use crate::app::RegistryApp;
 use crate::extractors::Token;
-use crate::network::registry::errors::RegistryError;
+use crate::registry::errors::RegistryError;
 use crate::types::RepositoryName;
 use actix_web::http::StatusCode;
 use actix_web::web::{Data, Path};
